2 command without test
“opme” command via cli style
“build-npm” command via cli style
both of  without test
