--- bin/utils/open-devtools.js
@@ -1,3 +1,0 @@
-const { open } =require('./cli-style');
-
-open();
