--- bin/utils/wrap.js
@@ -1,3 +1,0 @@
-module.exports = ( func, argv=process.argv )=>{
-  func(...argv)
-}
