+++ package.json
@@ -1,18 +1,29 @@
 {
   "name": "@mcswift/wechat-devtools-cli",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "main": "index.js",
   "repository": "https://github.com/mcTree/wechat-devtools-cli.git",
   "author": "McSwift <publish@mcswift.pro>",
   "license": "MIT",
+  "scripts": {
+    "build":"tsc"
+  },
   "bin": {
     "wx-devtools": "bin/devtools.js"
   },
   "dependencies": {
     "axios": "^0.19.2",
-    "fs-extra": "^9.0.0"
+    "chokidar": "^3.4.0",
+    "execa": "^4.0.2",
+    "fs-extra": "^9.0.0",
+    "lodash": "^4.17.15"
   },
   "publishConfig": {
     "access": "public"
+  },
+  "devDependencies": {
+    "@types/lodash": "^4.14.153",
+    "@types/node": "^14.0.5",
+    "typescript": "^3.9.3"
   }
 }
