+++ src/lib/command.ts
@@ -1,0 +1,43 @@
+import { getProjectDir,getMiniprogramRoot } from "./utils";
+import path from "path";
+import chokidar from "chokidar"
+
+
+export const open:CommandFunc = async (execute:ModeExecutor) => {
+  // http://127.0.0.1:端口号/v2/open?project=项目全路径
+  // cli open
+  return execute("open", "--project", getProjectDir());
+}
+
+export const buildNpm:CommandFunc = async (executor:ModeExecutor, options) => {
+  const projectDir = getProjectDir();
+  const cliParams:[Command,string,string] = ["build-npm", "--project", projectDir]
+  /**
+   * --watch 监听
+   * --project / --package 监听 package.json @default
+   * --lib / --modules 监听 node_modules
+   */
+  if (options.includes("--watch")) {
+    const miniprogramRoot = getMiniprogramRoot();
+    let watchPath = "";
+    const isLib = options.includes("--modules") || options.includes("--lib");
+    if (isLib) {
+      watchPath = path.join(projectDir, miniprogramRoot, "node_modules");
+    } else {
+      watchPath = path.resolve(projectDir, miniprogramRoot, "package.json");
+    }
+
+    return new Promise((resolve, reject) => {
+      try {
+        chokidar.watch(watchPath).on("all", (event, path) => {
+          executor(...cliParams)
+          resolve({ event, path });
+        });
+      } catch (e) {
+        reject(e);
+      }
+    });
+  }
+
+  return executor(...cliParams)
+}
