+++ src/lib/lanucher.ts
@@ -1,0 +1,6 @@
+export default function (
+  func: Function,
+  argv: Array<any> = process.argv
+): Promise<any> {
+  return func(...argv);
+};
