+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "compilerOptions": {
+    "moduleResolution":"node",
+    "target": "ESNext",
+    "sourceMap": true,
+    "outDir": "./dist",
+    "module": "CommonJS",
+    "allowSyntheticDefaultImports":true
+  },
+  "include": ["src/**/*"],
+}
